+++ src/message/mod.rs
@@ -12,5 +12,10 @@ pub mod builder;
 /// Security List Request and Security List messages
 pub mod security_list;
 
+/// Market data messages
+pub mod market_data;
+
+pub use admin::*;
 pub use builder::*;
+pub use market_data::*;
 pub use security_list::*;